perf: eliminate data fetching waterfalls and add dynamic imports
Performance optimizations for Analytics, Chat, and Insights pages:

- Convert Analytics, Chat, and Insights pages to server components that
  fetch organization data server-side, eliminating client-side waterfall
  where pages fetched org first, then data using org.id
- Extract client logic into separate *-content.tsx files that receive
  organization as props
- Add dynamic imports for Insights tab components (InsightsSummary,
  TopicTrends, MeetingPatterns, etc.) - only loads when tab is active
- Add dynamic imports for VideoPlayer component in video-player-with-progress
  and share page - reduces initial bundle size

These changes improve perceived performance by:
1. Removing sequential fetch patterns (org -> data waterfall)
2. Reducing initial JS bundle with code splitting
3. Loading heavy components on demand
